@@ -4,9 +4,9 @@
   "description": "react component pull to refresh navigation",
   "main": "index.js",
   "scripts": {
+    "start": "node server",
     "prepare": "npm run build",
     "build": "webpack",
-    "start": "cd test && node index.js",
     "build:test": "cd test && webpack --config webpack.config.release.js",
     "lint": "eslint src test",
     "lint:fix": "eslint src test --fix",
@@ -30,31 +30,29 @@
   "homepage": "https://github.com/yusukeshibata/react-pullrefresh#readme",
   "peerDependencies": {
     "react": "^15.4.2",
-    "react-native-svg": "^5.1.5",
-    "react-native": "^0.42.0"
+    "react-native": "^0.42.3",
+    "react-native-svg": "^5.1.6"
   },
   "devDependencies": {
-    "babel-core": "^6.23.1",
-    "babel-eslint": "^7.1.1",
-    "babel-loader": "^6.3.2",
+    "babel-core": "^6.24.0",
+    "babel-eslint": "^7.2.1",
+    "babel-loader": "^6.4.1",
     "babel-plugin-transform-object-assign": "^6.22.0",
     "babel-polyfill": "^6.23.0",
-    "babel-preset-es2015": "^6.22.0",
-    "babel-preset-latest": "^6.22.0",
+    "babel-preset-es2015": "^6.24.0",
+    "babel-preset-latest": "^6.24.0",
     "babel-preset-react": "^6.23.0",
     "babel-preset-stage-1": "^6.22.0",
-    "css-loader": "^0.26.1",
-    "eslint": "^3.16.1",
-    "eslint-loader": "^1.6.3",
-    "file-loader": "^0.10.0",
+    "css-loader": "^0.27.3",
+    "eslint": "^3.18.0",
+    "eslint-loader": "^1.7.0",
+    "file-loader": "^0.10.1",
     "react": "^15.4.2",
     "react-dom": "^15.4.2",
     "react-hot-loader": "^3.0.0-beta.6",
-    "react-native": "^0.42.0",
-    "react-native-svg": "^5.1.5",
-    "style-loader": "^0.13.2",
-    "url-loader": "^0.5.7",
-    "webpack": "1.14.0",
-    "webpack-dev-server": "1.16.2"
+    "style-loader": "^0.16.1",
+    "url-loader": "^0.5.8",
+    "webpack": "2.3.2",
+    "webpack-dev-server": "2.4.2"
   }
 }
