@@ -1,21 +1,21 @@
 import React, { Component } from 'react'
 
-export class div extends Component {
+export class Div extends Component {
   render() {
     return (<div {...this.props}/>)
   }
 }
-export class svg extends Component {
+export class Svg extends Component {
   render() {
     return (<svg {...this.props}/>)
   }
 }
-export class path extends Component {
+export class Path extends Component {
   render() {
     return (<path {...this.props}/>)
   }
 }
-export class circle extends Component {
+export class Circle extends Component {
   render() {
     return (<circle {...this.props}/>)
   }
